move OwnerCounts to separate file
